No commas operation

+++ src/main/java/fr/hetic/Calculator.java
@@ -3,11 +3,11 @@ package fr.hetic;
 class Calculator {
     public static void main(String[] args) {
         try {
-            Double number1 = Double.parseDouble(args[0]);
-            Double number2 = Double.parseDouble(args[1]);
+            Integer number1 = Integer.parseInt(args[0]);
+            Integer number2 = Integer.parseInt(args[1]);
             String operator = args[2];
 
-            Double result = calculate(number1, number2, operator);
+            Integer result = calculate(number1, number2, operator);
 
             System.out.println("Result: " + result);
         } catch (NumberFormatException e) {
@@ -17,7 +17,7 @@ class Calculator {
         }
     }
 
-    private static Double calculate(Double number1, Double number2, String operator) {
+    private static Integer calculate(Integer number1, Integer number2, String operator) {
         return switch (operator) {
             case "+" -> number1 + number2;
             case "-" -> number1 - number2;
